+++ acvm/src/pwg.rs
@@ -1,19 +1,23 @@
 // Re-usable methods that backends can use to implement their PWG
 
-use crate::{OpcodeNotSolvable, OpcodeResolutionError};
+use crate::{OpcodeNotSolvable, OpcodeResolutionError, PartialWitnessGenerator};
 use acir::{
+    circuit::opcodes::{Opcode, OracleData},
     native_types::{Expression, Witness},
     FieldElement,
 };
 use std::collections::BTreeMap;
 
-use self::arithmetic::ArithmeticSolver;
+use self::{
+    arithmetic::ArithmeticSolver, block::Blocks, directives::solve_directives, oracle::OracleSolver,
+};
 
 // arithmetic
 pub mod arithmetic;
 // Directives
 pub mod directives;
 // black box functions
+mod blackbox;
 pub mod block;
 pub mod hash;
 pub mod logic;
@@ -21,6 +25,109 @@ pub mod oracle;
 pub mod range;
 pub mod signature;
 pub mod sorting;
+
+#[derive(Debug, PartialEq)]
+pub enum PartialWitnessGeneratorStatus {
+    /// All opcodes have been solved.
+    Solved,
+
+    /// The `PartialWitnessGenerator` has encountered a request for [oracle data][Opcode::Oracle].
+    ///
+    /// The caller must resolve these opcodes externally and insert the results into the intermediate witness.
+    /// Once this is done, the `PartialWitnessGenerator` can be restarted to solve the remaining opcodes.
+    RequiresOracleData { required_oracle_data: Vec<OracleData>, unsolved_opcodes: Vec<Opcode> },
+}
+
+#[derive(Debug, PartialEq)]
+pub enum OpcodeResolution {
+    /// The opcode is resolved
+    Solved,
+    /// The opcode is not solvable
+    Stalled(OpcodeNotSolvable),
+    /// The opcode is not solvable but could resolved some witness
+    InProgress,
+}
+
+pub fn solve(
+    backend: &impl PartialWitnessGenerator,
+    initial_witness: &mut BTreeMap<Witness, FieldElement>,
+    blocks: &mut Blocks,
+    mut opcode_to_solve: Vec<Opcode>,
+) -> Result<PartialWitnessGeneratorStatus, OpcodeResolutionError> {
+    let mut unresolved_opcodes: Vec<Opcode> = Vec::new();
+    let mut unresolved_oracles: Vec<OracleData> = Vec::new();
+    while !opcode_to_solve.is_empty() || !unresolved_oracles.is_empty() {
+        unresolved_opcodes.clear();
+        let mut stalled = true;
+        let mut opcode_not_solvable = None;
+        for opcode in &opcode_to_solve {
+            let mut solved_oracle_data = None;
+            let resolution = match opcode {
+                Opcode::Arithmetic(expr) => ArithmeticSolver::solve(initial_witness, expr),
+                Opcode::BlackBoxFuncCall(bb_func) => {
+                    blackbox::solve(backend, initial_witness, bb_func)
+                }
+                Opcode::Directive(directive) => solve_directives(initial_witness, directive),
+                Opcode::Block(block) | Opcode::ROM(block) | Opcode::RAM(block) => {
+                    blocks.solve(block.id, &block.trace, initial_witness)
+                }
+                Opcode::Oracle(data) => {
+                    let mut data_clone = data.clone();
+                    let result = OracleSolver::solve(initial_witness, &mut data_clone)?;
+                    solved_oracle_data = Some(data_clone);
+                    Ok(result)
+                }
+            };
+            match resolution {
+                Ok(OpcodeResolution::Solved) => {
+                    stalled = false;
+                }
+                Ok(OpcodeResolution::InProgress) => {
+                    stalled = false;
+                    // InProgress Oracles must be externally re-solved
+                    if let Some(oracle) = solved_oracle_data {
+                        unresolved_oracles.push(oracle);
+                    } else {
+                        unresolved_opcodes.push(opcode.clone());
+                    }
+                }
+                Ok(OpcodeResolution::Stalled(not_solvable)) => {
+                    if opcode_not_solvable.is_none() {
+                        // we keep track of the first unsolvable opcode
+                        opcode_not_solvable = Some(not_solvable);
+                    }
+                    // We push those opcodes not solvable to the back as
+                    // it could be because the opcodes are out of order, i.e. this assignment
+                    // relies on a later opcodes' results
+                    unresolved_opcodes.push(match solved_oracle_data {
+                        Some(oracle_data) => Opcode::Oracle(oracle_data),
+                        None => opcode.clone(),
+                    });
+                }
+                Err(OpcodeResolutionError::OpcodeNotSolvable(_)) => {
+                    unreachable!("ICE - Result should have been converted to GateResolution")
+                }
+                Err(err) => return Err(err),
+            }
+        }
+        // We have oracles that must be externally resolved
+        if !unresolved_oracles.is_empty() {
+            return Ok(PartialWitnessGeneratorStatus::RequiresOracleData {
+                required_oracle_data: unresolved_oracles,
+                unsolved_opcodes: unresolved_opcodes,
+            });
+        }
+        // We are stalled because of an opcode being bad
+        if stalled && !unresolved_opcodes.is_empty() {
+            return Err(OpcodeResolutionError::OpcodeNotSolvable(
+                opcode_not_solvable
+                    .expect("infallible: cannot be stalled and None at the same time"),
+            ));
+        }
+        std::mem::swap(&mut opcode_to_solve, &mut unresolved_opcodes);
+    }
+    Ok(PartialWitnessGeneratorStatus::Solved)
+}
 
 // Returns the concrete value for a particular witness
 // If the witness has no assignment, then